Added a numeric benchmark base
This removes some of the bootstrapping code that is shared by all the benchmarks
