+++ src/Cistern.Linq.Benchmarking/Benchmarks/SumBenchmark.cs
@@ -6,24 +6,8 @@ using System.Text;
 namespace Cistern.Linq.Benchmarking.Benchmarks
 {
 	[CoreJob, MemoryDiagnoser]
-	public class SumBenchmark
+	public class SumBenchmark : NumericBenchmarkBase
 	{
-		[Params(1000, 1000000)]
-		public int NumberOfItems;
-
-		public double[] Numbers;
-
-		[GlobalSetup]
-		public void Setup()
-		{
-			Numbers = new double[NumberOfItems];
-			for (int i = 0; i < NumberOfItems; i++)
-			{
-				Numbers[i] = i;
-			}
-		}
-
-
 		[Benchmark]
 		public double ForLoop()
 		{
