+++ src/Cistern.Linq.Benchmarking/Benchmarks/ToListBenchmark.cs
@@ -6,23 +6,8 @@ using System.Text;
 namespace Cistern.Linq.Benchmarking.Benchmarks
 {
 	[CoreJob, MemoryDiagnoser]
-	public class ToListBenchmark
+	public class ToListBenchmark : NumericBenchmarkBase
 	{
-		[Params(10000, 1000000)]
-		public int NumberOfItems;
-
-		public double[] Numbers;
-
-		[GlobalSetup]
-		public void Setup()
-		{
-			Numbers = new double[NumberOfItems];
-			for (int i = 0; i < NumberOfItems; i++)
-			{
-				Numbers[i] = i + 1;
-			}
-		}
-
 		[Benchmark]
 		public List<double> ForLoop()
 		{
